+++ openfisca_core/variables/helpers.py
@@ -1,9 +1,11 @@
 from __future__ import annotations
 
-import sortedcontainers
+from openfisca_core.periods.typing import Period
 from typing import Optional
 
-from openfisca_core.periods import Period
+import sortedcontainers
+
+from openfisca_core import periods
 
 from .. import variables
 
@@ -18,7 +20,7 @@ def get_annualized_variable(variable: variables.Variable, annualization_period: 
 
         def annual_formula(population, period, parameters):
             if period.start.month != 1 and (annualization_period is None or period not in annualization_period):
-                return population(variable.name, period.this(YEAR).first_month)
+                return population(variable.name, period.this(periods.YEAR).this(periods.MONTH))
             if original_formula.__code__.co_argcount == 2:
                 return original_formula(population, period)
             return original_formula(population, period, parameters)